WIP: Permutation synthesis
